Dejar controllers bien

+++ PROG/pruebaproyectoBrunoDelgado/src/main/java/com/salesianostriana/dam/pruebaproyecto/controller/LotesController.java
@@ -11,6 +11,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import com.salesianostriana.dam.pruebaproyecto.model.Lote;
+import com.salesianostriana.dam.pruebaproyecto.service.LineaDePedidoService;
 import com.salesianostriana.dam.pruebaproyecto.service.LoteService;
 import com.salesianostriana.dam.pruebaproyecto.service.ProdUnidadService;
 
@@ -22,7 +23,10 @@ public class LotesController {
 	
 	@Autowired
 	private ProdUnidadService servicioProdUnidad;
-
+	
+	@Autowired
+	private LineaDePedidoService servicioLdp;
+	
 	@GetMapping("/lotes")
 	public String controladorLotes(Model model) {
 		model.addAttribute("listaLotes", servicioLote.findAll());
@@ -71,7 +75,17 @@ public class LotesController {
 	
 	@GetMapping("/borrarLote/{id}")
 	public String borrar(@PathVariable("id") long id) {
-		//servicio.deleteById(id);
+		if(servicioLdp.buscarProductoEnLp(id).isPresent()) {
+			return "noBorrar";
+		}
+		
+		
+		
+		
+		servicioLdp.deleteById(id);
+		
+
+
 		servicioProdUnidad.deleteByIdConFavoritos(id);
 		return "redirect:/admin/lotes/listaLotes";
 	}
